+++ ecs.php
@@ -34,7 +34,7 @@ return ECSConfig::configure()
                 'construct',
                 'destruct',
                 'magic',
-                'phpunit',
+                'method_protected_abstract',
                 'method_public',
                 'method_protected',
                 'method_private',